ai-day1/fastapi_face_pipeline/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
mediapipe>=0.10.8,<0.11
opencv-python>=4.9.0.80
numpy>=1.26
python-multipart>=0.0.9
Pillow>=10.2
supabase>=2.4.0
python-dotenv>=1.0.1
